New ports created on component will now be placed correctly on instance
When new ports are created on a layout component, the layout instance
will receive signals about this. In response to these signals the
instance will now try to add new ports near the position where they were
added on the component. If the edge has no room, other edges will be
used instead. If none of the edges have any room, the instance will be
resized vertically to make room.