Don't parse and rewrite architectures for non-graphical VHDL files
The program now distinguishes between files opened in graphical mode and
files opened in text mode. The former are meant to be managed by the
program itself and will not contain much more than an entity and an
architecture with signals and component instantiations. The latter are
meant to be edited by the user and can contain whatever the user wants
in the architecture. VHDE should only care about the interface (entity)
of such manually edited files.