Always save ports in layout files in the order they were added
Also swap the name and the position for ports in layout files.

An important goal of VHDE's text files for saving layout is to make
diffs understandable by humans. By maintaining the creation order of
ports, renaming ports or rearranging them on a diagram will only change
the lines of the affected ports and thereby minimize the diff.