Move responsibility for associating added ports out of GuiInstance
GuiInstance only exists when the instance is in the currently open
diagram, so this class does not exist when ports are added to a
component in a different diagram. Instead the Project::PortConnector
class was added with as only purpose to wait for the port_added events
of the LayoutPort and the VHDLPort and then associate the latter with
the former.