Fix crash when ports are added on component with multiple instances
PortConnector will not always receive the notification for to be
connected VHDL and Layout ports after each other. It may get multiple
VHDL port notifications after each other when a port is added on a
component that has multiple instances in an architecture. Therefore it
needs to create a list of still to be connected ports and connect each
VHDL port to a Layout port on a first-come first-served basis.